+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.7.0
 FEATURES:
-- Added `snowflake` input in `trocco_job_definition` resource.
+- Added `snowflake` input/output in `trocco_job_definition` resource.
 - Added `s3` type for `trocco_connection` resource.
 
 ## 0.6.0